Add files via upload
In view of the traditional methods of high-resolution remote sensing image semantic segmentation, there are slow calculation speeds, low efficiency, large data sets, and difficult to deal with remote sensing image-specific road dense fork roads, small gaps in buildings and other details, this paper proposes a remote sensing image semantics The segmented loss function and its network structure. In order to avoid the problem of gradient disappearance caused by deep networks, based on the U-Net structure, combined with ResNet to form a Res-Unet network, while retaining a clear U-shaped encoding and decoding structure, it incorporates the residual module of ResNet. The initial setting is He initialization and the introduction of a new optimizer Stochastic Weight Averaging to help learning convergence. At the same time, in order to better solve the problems of general learning effect of small data sets and difficulty in learning convergence in remote sensing image semantic segmentation tasks, this paper designs a cross-entropy loss function that can effectively maintain the stable convergence of network learning and can significantly improve the network's mIoU score The Lovasz Softmax loss function is combined to form a new loss. This paper conducts a control experiment on the competition data set of "BDCI 2017 Jiage Data Satellite Image AI Classification and Recognition", and compares the performance of image segmentation with three loss functions of cross entropy (cross entropy), tversky, and Lovasz. In the experiment, the Lovasz-Crossentropy Softmax loss mIoU score is as high as 0.84, which obviously exceeds the cross entropy of 0.79 and Lovasz's 0.81; f1-score reaches 0.91, which is also better than the cross entropy of 0.87 and Lovasz's 0.89.